requirements.txt
fastapi
tortoise-orm
numpy
uvicorn
Jinja2
websockets